feat: Add support for deployment diagrams (#8)
This commit adds the new DeploymentNode element type as well as the
supporting Property type in order to allow the creation of deployment
diagrams (https://c4model.com/#DeploymentDiagram).

Notably, the structure for DeploymentNode differs from the previous
element types in that the creation args take a slice of elements as a
parameter, allowing you to skip additional calls to AddElement provided
you create your elements in order from contained to container. This will
likely serve as the template for upcoming improvements to the API for
all elements.