feat: Add an option for LayoutLandscape
This commit adds a new lanscape option for layout which is much easier
to use than the PlantUML native left to right layout.

Since the LAYOUT_LEFT_RIGHT algorithm is so stupid as to be pointless,
this commit also deprecates the existing LayoutLeftRight option to pave
the way for replumbing that in the next major version.